Flipped driver controls based on driver station
Enabling limelight support highlighted a fundamental issue with driving.
When driving from the red driver station, the robot would flip its
controls when it picked up on the AprilTags on the other end of the
field.
There were two problems at work here. The Limelight was incorrectly
configured to think it was at the front of the robot when it technically
was at the back. This needs to be fixed in the Limelight configuration.
This meant the robot thought it was 180° from its actual field-oriented
rotation, which meant the controls behaved correctly, until coming up
against the AprilTags from the other end of the field.
This is making my brain hurt, but based on the field-relative positions
and assuming the Limelight isn't doing something weird, this makes sense
and should help to address the issue by ensuring the robot is behaving
properly within the field coordinate space.